Fix v2 prise en compte des ante

+++ src/main/java/fr/eurekapoker/parties/domaine/poker/actions/ActionPokerAvecBet.java
@@ -7,7 +7,7 @@ public class ActionPokerAvecBet extends ActionPokerJoueur {
     // todo gérer le paramètre TotalBet
     private final BigDecimal montantAction;
     private final boolean totalBet;
-    public ActionPokerAvecBet(String nomJoueur, TypeAction typeAction, float montantAction, boolean totalBet) {
+    public ActionPokerAvecBet(String nomJoueur, TypeAction typeAction, double montantAction, boolean totalBet) {
         super(nomJoueur, typeAction);
         this.montantAction = new BigDecimal(montantAction).setScale(2, RoundingMode.HALF_UP);
         this.totalBet = totalBet;
